# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CardLink {
  /** Link type: whatsapp, phone, sms, email, linkedin, facebook, instagram, twitter, youtube, github, custom */
  type: string;
  /** @nullable */
  label?: string | null;
  url: string;
  /**
     * Emoji or icon name for custom links
     * @nullable
     */
  icon?: string | null;
  sortOrder?: number;
}

export interface Card {
  id: number;
  userId: string;
  /** @nullable */
  username?: string | null;
  isPublished: boolean;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandSubtitle?: string | null;
  fullName: string;
  /** @nullable */
  professionalTitle?: string | null;
  /** @nullable */
  location?: string | null;
  showStatusDot: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** CSS hex color */
  accentColor: string;
  textColor: string;
  bgColor: string;
  /** One of: outfit, playfair, inter */
  fontStyle: string;
  /** One of: tech, real-estate, creative, medical */
  bgIconPack: string;
  bgIconDensity: number;
  bgIconOpacity: number;
  links: CardLink[];
  createdAt: string;
  updatedAt: string;
}

export interface CardInput {
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandSubtitle?: string | null;
  fullName: string;
  /** @nullable */
  professionalTitle?: string | null;
  /** @nullable */
  location?: string | null;
  showStatusDot?: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  accentColor?: string;
  textColor?: string;
  bgColor?: string;
  fontStyle?: string;
  bgIconPack?: string;
  bgIconDensity?: number;
  bgIconOpacity?: number;
  links?: CardLink[];
}

export interface CardUpdate {
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandSubtitle?: string | null;
  fullName?: string;
  /** @nullable */
  professionalTitle?: string | null;
  /** @nullable */
  location?: string | null;
  showStatusDot?: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  accentColor?: string;
  textColor?: string;
  bgColor?: string;
  fontStyle?: string;
  bgIconPack?: string;
  bgIconDensity?: number;
  bgIconOpacity?: number;
  links?: CardLink[];
}

export interface PublishInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
}

export interface QrCode {
  /** SVG string of the QR code */
  svg: string;
  /** The public URL the QR code points to */
  url: string;
}

export interface HtmlExport {
  /** Complete standalone HTML file content */
  html: string;
  filename: string;
}

export interface UsernameCheck {
  available: boolean;
  username: string;
}

export type CheckUsernameParams = {
username: string;
};