# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CardLink } from './cardLink';

export interface Card {
  id: number;
  userId: string;
  /** @nullable */
  username?: string | null;
  isPublished: boolean;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandSubtitle?: string | null;
  fullName: string;
  /** @nullable */
  professionalTitle?: string | null;
  /** @nullable */
  location?: string | null;
  showStatusDot: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** CSS hex color */
  accentColor: string;
  textColor: string;
  bgColor: string;
  /** One of: outfit, playfair, inter */
  fontStyle: string;
  /** One of: tech, real-estate, creative, medical */
  bgIconPack: string;
  bgIconDensity: number;
  bgIconOpacity: number;
  links: CardLink[];
  createdAt: string;
  updatedAt: string;
}