# lib/api-zod/src/generated/types/cardLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CardLink {
  /** Link type: whatsapp, phone, sms, email, linkedin, facebook, instagram, twitter, youtube, github, custom */
  type: string;
  /** @nullable */
  label?: string | null;
  url: string;
  /**
     * Emoji or icon name for custom links
     * @nullable
     */
  icon?: string | null;
  sortOrder?: number;
}