# lib/api-zod/src/generated/types/cardUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CardLink } from './cardLink';

export interface CardUpdate {
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandSubtitle?: string | null;
  fullName?: string;
  /** @nullable */
  professionalTitle?: string | null;
  /** @nullable */
  location?: string | null;
  showStatusDot?: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  accentColor?: string;
  textColor?: string;
  bgColor?: string;
  fontStyle?: string;
  bgIconPack?: string;
  bgIconDensity?: number;
  bgIconOpacity?: number;
  links?: CardLink[];
}