# lib/api-zod/src/generated/types/htmlExport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HtmlExport {
  /** Complete standalone HTML file content */
  html: string;
  filename: string;
}