# lib/api-zod/src/generated/types/publishInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PublishInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
}