# lib/api-zod/src/generated/types/qrCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Business Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QrCode {
  /** SVG string of the QR code */
  svg: string;
  /** The public URL the QR code points to */
  url: string;
}